Move each model to a separate file
